pyproject.toml
[tool.poetry]
name = "imgextract"
version = "0.1.0"
description = ""
authors = ["Linh Bui <linh.cn3@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.0.3"
opencv-python = "^4.10.0.84"
pillow = "^10.4.0"
tensorflow = "^2.17.0"
matplotlib = "^3.9.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.scripts]
start = "imgextract.main:main"